+++ Ecommerce-backend/transaction/src/main/java/com/ecommerce/transaction/service/impl/TransactionServiceImpl.java
@@ -1,8 +1,13 @@
 package com.ecommerce.transaction.service.impl;
 
 import com.ecommerce.transaction.entity.Transaction;
+import com.ecommerce.transaction.exception.TransactionAlreadyExistsException;
+import com.ecommerce.transaction.exception.TransactionNotFoundException;
 import com.ecommerce.transaction.repository.TransactionRepository;
 import com.ecommerce.transaction.service.TransactionService;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -32,8 +37,8 @@ public class TransactionServiceImpl implements TransactionService {
 
     @Override
     public Transaction getTransactionById(String transactionId) {
-        Optional<Transaction> optionalTransaction = transactionRepository.findById(transactionId);
-        return optionalTransaction.orElse(null);
+        Transaction optionalTransaction = transactionRepository.findById(transactionId).orElseThrow(() -> new TransactionNotFoundException("Transaction not found with ID: " + transactionId));
+        return optionalTransaction;
     }
 
     @Override
@@ -43,7 +48,7 @@ public class TransactionServiceImpl implements TransactionService {
         else {
             Optional<Transaction> existingTransaction = transactionRepository.findById(transaction.getTransactionId());
             if (existingTransaction.isPresent()) {
-                throw new RuntimeException("Transaction with ID " + transaction.getTransactionId() + " already exists.");
+                throw new TransactionAlreadyExistsException("Transaction with ID " + transaction.getTransactionId() + " already exists.");
             }
         }
         return transactionRepository.saveAndFlush(transaction);
@@ -51,40 +56,34 @@ public class TransactionServiceImpl implements TransactionService {
 
     @Override
     public Transaction updateTransaction(Transaction updatedTransaction) {
+        Optional<Transaction> existingTransaction = transactionRepository.findById(updatedTransaction.getTransactionId());
+        if (existingTransaction.isEmpty()) {
+            throw new TransactionNotFoundException("Transaction with ID " + updatedTransaction.getTransactionId() + " not found.");
+        }
         return transactionRepository.saveAndFlush(updatedTransaction);
     }
 
     @Override
     public void patchTransaction(String transactionId, Map<String, Object> updates) {
         Transaction existingTransaction = transactionRepository.findById(transactionId)
-                .orElseThrow(() -> new RuntimeException("Transaction not found"));
+                .orElseThrow(() -> new TransactionNotFoundException("Transaction not found with ID: " + transactionId));
 
-        updates.forEach((key, value) -> {
-            try {
-                Field field = Transaction.class.getDeclaredField(key);
-                field.setAccessible(true);
-                if (field.getType() == Timestamp.class) {
-                    try {
-                        OffsetDateTime odt = OffsetDateTime.parse((String) value);
-                        Timestamp timestampValue = Timestamp.from(odt.toInstant());
-                        field.set(existingTransaction, timestampValue);
-                    } catch (DateTimeParseException e) {
-                        throw new IllegalArgumentException("Invalid format for " + key + " TimeStamp field");
-                    }
-                }
-                else {
-                    field.set(existingTransaction, value);
-                }
-            } catch (NoSuchFieldException | IllegalAccessException e) {
-                throw new IllegalArgumentException("Invalid update field: " + key);
-            }
-        });
+        ObjectMapper mapper = new ObjectMapper();
+        mapper.registerModule(new JavaTimeModule()); // for Timestamp support
+
+        try {
+            // assuming updates is a Map<String, Object>
+            mapper.updateValue(existingTransaction, updates);
+        } catch (JsonProcessingException e) {
+            throw new IllegalArgumentException("Invalid update field: " + updates);
+        }
 
         transactionRepository.save(existingTransaction);
     }
 
     @Override
     public void deleteTransaction(String transactionId) {
+        if(!transactionRepository.existsById(transactionId)) throw new TransactionNotFoundException("Transaction not found with ID: " + transactionId);
         transactionRepository.deleteById(transactionId);
     }
 }
